chore: move service to /domain folder
