+++ src/main/java/lems/cowshed/api/controller/user/UserApiController.java
@@ -11,9 +11,9 @@ import lems.cowshed.domain.mail.code.CodeFinder;
 import lems.cowshed.domain.mail.code.CodeType;
 import lems.cowshed.domain.user.User;
 import lems.cowshed.exception.BusinessException;
-import lems.cowshed.service.CustomUserDetails;
+import lems.cowshed.domain.user.CustomUserDetails;
 import lems.cowshed.service.MailService;
-import lems.cowshed.service.UserService;
+import lems.cowshed.service.user.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
