+++ src/main/java/lems/cowshed/domain/user/CustomUserDetails.java
@@ -1,12 +1,10 @@
-package lems.cowshed.service;
+package lems.cowshed.domain.user;
 
-import lems.cowshed.domain.user.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
