+++ src/test/java/lems/cowshed/api/controller/bookmark/BookmarkApiControllerTest.java
@@ -1,13 +1,7 @@
 package lems.cowshed.api.controller.bookmark;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import lems.cowshed.domain.user.Role;
-import lems.cowshed.domain.user.User;
 import lems.cowshed.service.BookmarkService;
-import lems.cowshed.service.CustomUserDetails;
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -15,9 +9,6 @@ import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.MockMvc;
 
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.user;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 @WithMockUser
 @WebMvcTest(controllers = BookmarkApiController.class)
