+++ src/test/java/lems/cowshed/domain/BaseEntityTest.java
@@ -7,7 +7,7 @@ import lems.cowshed.domain.user.Mbti;
 import lems.cowshed.domain.user.User;
 import lems.cowshed.domain.user.UserRepository;
 import lems.cowshed.domain.userevent.UserEventRepository;
-import lems.cowshed.service.UserService;
+import lems.cowshed.service.user.UserService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
